# Conflicts:
#	Cargo.lock
#	core/vpn/src/network.rs

+++ exe-unit/src/state.rs
@@ -141,7 +141,7 @@ impl Batch {
                     0 => CommandResult::Ok,
                     _ => CommandResult::Error,
                 });
-                self.notifier.notify(idx as usize);
+                self.notifier.notify(idx);
                 Some(event)
             }
             RuntimeEventKind::StdOut(out) => {
